Extract slider metric presentation logic
