+++ packages/twenty-front/src/modules/ai/components/AiModelTierSlider.tsx
@@ -7,12 +7,6 @@ import {
   isAiModelEffort,
 } from 'twenty-shared/ai';
 import { isDefined } from 'twenty-shared/utils';
-import {
-  IconBolt,
-  IconBrain,
-  IconCoins,
-  type IconComponent,
-} from 'twenty-ui/icon';
 import { AppTooltip, TooltipDelay } from 'twenty-ui/surfaces';
 import { ThemeContext, themeCssVariables } from 'twenty-ui/theme-constants';
 
@@ -21,7 +15,7 @@ import { getAiModelEffortLabel } from '@/ai/utils/getAiModelEffortLabel';
 import { formatMetricDelta } from '@/ai/utils/formatMetricDelta';
 import { getAiModelModeDescription } from '@/settings/ai/utils/getAiModelModeDescription';
 import { getModelIcon } from '@/settings/ai/utils/getModelIcon';
-import { formatNumber } from '~/utils/format/formatNumber';
+import { getAiModelTierMetrics } from '@/ai/utils/getAiModelTierMetrics';
 
 const TRACK_HEIGHT_PX = 24;
 const TRACK_INSET_PX = 0;
@@ -29,7 +23,6 @@ const DOT_CENTER_INSET_PX = 18;
 const HANDLE_WIDTH_PX = 12;
 const HANDLE_HEIGHT_PX = 28;
 const LAST_STEP = AI_MODEL_TIERS.length - 1;
-const BALANCED_STEP = AI_MODEL_TIERS.indexOf('balanced');
 
 const StyledContainer = styled.div`
   display: flex;
@@ -161,29 +154,6 @@ const StyledInput = styled.input`
   width: 100%;
 `;
 
-type TierMetric = {
-  key: string;
-  Icon: IconComponent;
-  deltaPercent: number | undefined;
-  tooltipTitle?: string;
-  description: string;
-};
-
-const getComparison = (
-  delta: number,
-  labels: { multiplier: string; lower: string; higher: string; equal: string },
-) => {
-  if (delta >= 100) return labels.multiplier;
-  if (delta < 0) return labels.lower;
-  if (delta === 0) return labels.equal;
-  return labels.higher;
-};
-
-const hasDeltaToShow = (
-  metric: TierMetric,
-): metric is TierMetric & { deltaPercent: number } =>
-  isDefined(metric.deltaPercent);
-
 type AiModelTierSliderProps = {
   selectedTier: AiModelTier;
   onTierChange: (tier: AiModelTier) => void;
@@ -213,21 +183,6 @@ export const AiModelTierSlider = ({
     ? t` Not measured at this effort yet, so this is the base model's reading.`
     : '';
 
-  const intelligenceDelta = resolvedTier.intelligenceDeltaPercent ?? 0;
-  const intelligenceComparison = getComparison(intelligenceDelta, {
-    multiplier: t`${formatMetricDelta(intelligenceDelta)} the score of Balanced`,
-    lower: t`${formatNumber(Math.abs(intelligenceDelta))}% lower score than Balanced`,
-    higher: t`${formatNumber(intelligenceDelta)}% higher score than Balanced`,
-    equal: t`Same score as Balanced`,
-  });
-  const intelligenceScore = t`Intelligence score: ${formatNumber(model?.intelligenceIndex ?? 0)}`;
-  const costDelta = resolvedTier.costDeltaPercent ?? 0;
-  const costComparison = getComparison(costDelta, {
-    multiplier: t`${formatMetricDelta(costDelta)} the cost of Balanced.`,
-    lower: t`${formatNumber(Math.abs(costDelta))}% lower cost than Balanced.`,
-    higher: t`${formatNumber(costDelta)}% higher cost than Balanced.`,
-    equal: t`Same cost as Balanced.`,
-  });
   const modelEffort = model?.effort;
   const modelName = getAiModelModeDescription(resolvedTier, {
     showAutomatic: false,
@@ -238,46 +193,7 @@ export const AiModelTierSlider = ({
       ? getAiModelEffortLabel(modelEffort)
       : t`Default`;
 
-  // Below Balanced the gain is speed, above it intelligence; cost moves with
-  // both, so each side shows the two figures that explain the trade.
-  const candidateMetrics: TierMetric[] = [
-    ...(selectedStep < BALANCED_STEP
-      ? [
-          {
-            key: 'speed',
-            Icon: IconBolt,
-            deltaPercent: resolvedTier.speedDeltaPercent,
-            description:
-              (resolvedTier.speedDeltaPercent ?? 0) >= 100
-                ? t`${formatNumber(model?.outputTokensPerSecond ?? 0)} tokens/s, ${formatMetricDelta(resolvedTier.speedDeltaPercent ?? 0)} the speed of Balanced Mode.`
-                : t`${formatNumber(model?.outputTokensPerSecond ?? 0)} tokens/s, a ${formatMetricDelta(resolvedTier.speedDeltaPercent ?? 0)} change in speed compared with Balanced Mode.`,
-          },
-        ]
-      : []),
-    ...(selectedStep !== BALANCED_STEP
-      ? [
-          {
-            key: 'cost',
-            Icon: IconCoins,
-            deltaPercent: resolvedTier.costDeltaPercent,
-            description: costComparison,
-          },
-        ]
-      : []),
-    ...(selectedStep > BALANCED_STEP
-      ? [
-          {
-            key: 'intelligence',
-            Icon: IconBrain,
-            deltaPercent: resolvedTier.intelligenceDeltaPercent,
-            tooltipTitle: intelligenceScore,
-            description: intelligenceComparison,
-          },
-        ]
-      : []),
-  ];
-
-  const metrics = candidateMetrics.filter(hasDeltaToShow);
+  const metrics = getAiModelTierMetrics(resolvedTier);
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
     const tier = AI_MODEL_TIERS[Number(event.target.value)];
